adicionado metodo de ordenação

+++ src/com/ucs/modelos/Consulta.java
@@ -4,7 +4,7 @@ package com.ucs.modelos;
 import java.util.Date;
 
 public class Consulta {
-    public int ID;
+    public long ID;
     public String Medico;
     public String Paciente;
     public String Data;
